+++ crates/stackable-webhook/CHANGELOG.md
@@ -12,7 +12,8 @@ All notable changes to this project will be documented in this file.
   To achieve this, a new `CertificateResolver` was added.
   Also, `TlsServer::new` now returns an additional `mpsc::Receiver<Certificate>`, so that the caller
   can get notified about certificate rotations happening ([#1066]).
-- `stackable_webhook::Options` has been renamed to `stackable_webhook::WebhookOptions` ([#1066]).
+- `stackable_webhook::Options` has been renamed to `stackable_webhook::WebhookOptions`, as well as
+  `OptionsBuilder` to `WebhookOptionsBuilder` ([#1066]).
 
 ### Removed